Start adding label format option for users
