+++ app/controllers/observations/downloads_controller.rb
@@ -24,8 +24,7 @@ module Observations
     def print_labels
       query = find_query(:Observation)
       if query
-        # render_report(ObservationLabels.new(query))
-        render_report(RtfLabels.new(query))
+        render_report(ObservationLabels.new(@user, query))
       else
         flash_error(:runtime_search_has_expired.t)
         redirect_back_or_default("/")
@@ -49,8 +48,7 @@ module Observations
       elsif params[:commit] == :DOWNLOAD.l
         create_and_render_report
       elsif params[:commit] == :download_observations_print_labels.l
-        # render_report(ObservationLabels.new(@query))
-        render_report(RtfLabels.new(@query))
+        render_report(ObservationLabels.new(@user, @query))
       end
     end
